requirements.txt
gradio
langchain
langchain-experimental
langchain-openai
matplotlib
nltk
numpy
pandas
python-dotenv
tensorflow
weaviate-client
